# apps/web/openapi.json
{
  "components": {
    "schemas": {
...
      "ArchiveEntityDetail": {
        "description": "One entity, with tags and similarity expanded.\n\nDeliberately *not* a subclass of the summary. Its ``tags`` are objects with rank and\ncount where the summary's are strings, and narrowing a field's type in a subclass is\nnot a legal override -- mypy rejected it, which is the correct answer rather than an\ninconvenience. Two shapes that differ this much are two models.",
        "properties": {
          "extra": {
            "additionalProperties": true,
            "title": "Extra",
            "type": "object"
          },
          "first_seen_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "First Seen At"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "identity": {
            "title": "Identity",
            "type": "string"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "last_seen_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Seen At"
          },
          "latest_response_id": {
            "title": "Latest Response Id",
            "type": "string"
          },
          "listeners": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Listeners"
          },
          "mbid": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mbid"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "overview": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Overview"
          },
          "playcount": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Playcount"
          },
          "similar": {
            "items": {
              "$ref": "#/components/schemas/ArchiveSimilarityView"
            },
            "title": "Similar",
            "type": "array"
          },
          "tags": {
            "items": {
              "$ref": "#/components/schemas/ArchiveTagView"
            },
            "title": "Tags",
            "type": "array"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          }
        },
        "required": [
          "id",
          "identity",
          "kind",
          "name",
          "latest_response_id"
        ],
        "title": "ArchiveEntityDetail",
        "type": "object"
      },
      "ArchiveEntityListResponse": {
        "description": "A page of archived entities.\n\nThe field names match what the endpoint has always returned. Writing them from\nmemory instead of reading the code produced a model that rejected every real\nresponse -- which is precisely why these shapes are worth declaring.",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ArchiveEntitySummary"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "title": "Page Size",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "page_size"
        ],
        "title": "ArchiveEntityListResponse",
        "type": "object"
      },
      "ArchiveEntitySummary": {
        "description": "A list row: tags are bare names, because the list has no room for counts.",
        "properties": {
          "extra": {
            "additionalProperties": true,
            "title": "Extra",
            "type": "object"
          },
          "first_seen_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "First Seen At"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "identity": {
            "title": "Identity",
            "type": "string"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "last_seen_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Seen At"
          },
          "latest_response_id": {
            "title": "Latest Response Id",
            "type": "string"
          },
          "listeners": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Listeners"
          },
          "mbid": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mbid"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "overview": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Overview"
          },
          "playcount": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Playcount"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "title": "Tags",
            "type": "array"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          }
        },
        "required": [
          "id",
          "identity",
          "kind",
          "name",
          "latest_response_id"
        ],
        "title": "ArchiveEntitySummary",
        "type": "object"
      },
...
      "ArchiveSimilarityView": {
        "properties": {
          "match": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Match"
          },
          "mbid": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mbid"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "rank": {
            "title": "Rank",
            "type": "integer"
          }
        },
        "required": [
          "name",
          "rank"
        ],
        "title": "ArchiveSimilarityView",
        "type": "object"
      },
...
      "ArchiveTagCountView": {
        "properties": {
          "entity_count": {
            "title": "Entity Count",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "norm": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Norm"
          }
        },
        "required": [
          "name",
          "entity_count"
        ],
        "title": "ArchiveTagCountView",
        "type": "object"
      },
      "ArchiveTagListResponse": {
        "properties": {
          "tags": {
            "items": {
              "$ref": "#/components/schemas/ArchiveTagCountView"
            },
            "title": "Tags",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "tags",
          "total"
        ],
        "title": "ArchiveTagListResponse",
        "type": "object"
      },
      "ArchiveTagView": {
        "properties": {
          "count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Count"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "rank": {
            "title": "Rank",
            "type": "integer"
          }
        },
        "required": [
          "name",
          "rank"
        ],
        "title": "ArchiveTagView",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ReindexCounts": {
        "properties": {
          "albums": {
            "title": "Albums",
            "type": "integer"
          },
          "aliases": {
            "title": "Aliases",
            "type": "integer"
          },
          "artists": {
            "title": "Artists",
            "type": "integer"
          },
          "entity_tags": {
            "title": "Entity Tags",
            "type": "integer"
          },
          "expected_no_envelope": {
            "title": "Expected No Envelope",
            "type": "integer"
          },
          "observations": {
            "title": "Observations",
            "type": "integer"
          },
          "response_bodies": {
            "title": "Response Bodies",
            "type": "integer"
          },
          "similarities": {
            "title": "Similarities",
            "type": "integer"
          },
          "tag_edges": {
            "title": "Tag Edges",
            "type": "integer"
          },
          "tracks": {
            "title": "Tracks",
            "type": "integer"
          },
          "unexpected_shapes": {
            "title": "Unexpected Shapes",
            "type": "integer"
          }
        },
        "required": [
          "artists",
          "albums",
          "tracks",
          "tag_edges",
          "similarities",
          "aliases",
          "entity_tags",
          "observations",
          "response_bodies",
          "unexpected_shapes",
          "expected_no_envelope"
        ],
        "title": "ReindexCounts",
        "type": "object"
      },
      "ReindexRequest": {
        "properties": {
          "dry_run": {
            "default": true,
            "description": "Derive and report without writing. Defaults to true: a rebuild that writes should be an explicit choice, not an accident.",
            "title": "Dry Run",
            "type": "boolean"
          },
          "only": {
            "anyOf": [
              {
                "enum": [
                  "artist",
                  "album",
                  "track",
                  "tag",
                  "similar",
                  "alias"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Only"
          }
        },
        "title": "ReindexRequest",
        "type": "object"
      },
      "ReindexResponse": {
        "properties": {
          "changes": {
            "additionalProperties": true,
            "title": "Changes",
            "type": "object"
          },
          "counts": {
            "$ref": "#/components/schemas/ReindexCounts"
          },
          "dry_run": {
            "title": "Dry Run",
            "type": "boolean"
          },
          "duration_ms": {
            "title": "Duration Ms",
            "type": "integer"
          }
        },
        "required": [
          "dry_run",
          "counts",
          "duration_ms"
        ],
        "title": "ReindexResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Browse a Jellyfin music library, read artist/album/track metadata from the\nLast.fm API, keep a persistent incremental local archive of every Last.fm\nresponse, review a field-level diff, and apply the accepted changes back to\nJellyfin.\n",
    "title": "metaedit",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/archive/entities": {
      "get": {
        "description": "Search the derived layer offline.\n\nReads what the derivation produced and never touches the network: this is what\nanswers \"what do I already know about X\" without spending a request.",
        "operationId": "archive_entities_api_archive_entities_get",
        "parameters": [
          {
            "in": "query",
            "name": "kind",
            "required": false,
            "schema": {
              "default": "artist",
              "enum": [
                "artist",
                "album",
                "track"
              ],
              "title": "Kind",
              "type": "string"
            }
          },
          {
            "description": "only entities carrying this tag",
            "in": "query",
            "name": "tag",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "only entities carrying this tag",
              "title": "Tag"
            }
          },
          {
            "in": "query",
            "name": "search",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArchiveEntityListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Archive Entities",
        "tags": [
          "archive"
        ]
      }
    },
    "/api/archive/entities/{kind}/{entity_id}": {
      "get": {
        "operationId": "archive_entity_api_archive_entities__kind___entity_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "kind",
            "required": true,
            "schema": {
              "enum": [
                "artist",
                "album",
                "track"
              ],
              "title": "Kind",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "entity_id",
            "required": true,
            "schema": {
              "title": "Entity Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArchiveEntityDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Archive Entity",
        "tags": [
          "archive"
        ]
      }
    },
    "/api/archive/reindex": {
      "post": {
        "description": "Rebuild the derived layer from the raw archive.\n\n``dry_run`` defaults to **true** here, unlike the CLI: an HTTP client should not\nbe able to rebuild the structured layer by accident, and the report is what a UI\nwants in either case.",
        "operationId": "run_reindex_api_archive_reindex_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReindexRequest",
                "default": {
                  "dry_run": true
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReindexResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run Reindex",
        "tags": [
          "archive"
        ]
      }
    },
...
    "/api/archive/tags": {
      "get": {
        "description": "Every tag across the archive, with how many distinct entities carry it.",
        "operationId": "archive_tags_api_archive_tags_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 1000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArchiveTagListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Archive Tags",
        "tags": [
          "archive"
        ]
      }
    },
...
  }
}